@@ -1,12 +1,14 @@
-import Vue from 'vue'
-import App from './App.vue'
-import router from './router'
-import store from './store'
+import { createApp } from "vue";
+import App from "./App.vue";
+import router from "./router";
+import store from "./store";
 
-Vue.config.productionTip = false
+createApp(App).use(router).use(store).mount("#app");
 
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount('#app')
+// createApp.config.productionTip = false;
+
+// new createApp({
+//   router,
+//   store,
+//   render: (h) => h(App),
+// }).$mount("#app");
