@@ -1,27 +1,25 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import HomeView from '../views/HomeView.vue'
+import { createRouter, createWebHistory } from "vue-router";
 
-Vue.use(VueRouter)
+import HomeView from "@/views/HomeView.vue";
+import RegisterView from "@/views/RegisterView.vue";
+// import AboutView from "@/views/AboutView.vue";
 
 const routes = [
   {
-    path: '/',
-    name: 'home',
-    component: HomeView
+    path: "/",
+    name: "home",
+    component: HomeView,
   },
   {
-    path: '/about',
-    name: 'about',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/AboutView.vue')
-  }
-]
+    path: "/register",
+    name: "register",
+    component: RegisterView,
+  },
+];
 
-const router = new VueRouter({
-  routes
-})
+const router = new createRouter({
+  routes,
+  history: createWebHistory(),
+});
 
-export default router
+export default router;
