@@ -1,17 +1,11 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import { createStore } from "vuex";
+import auth from "@/store/modules/auth";
 
-Vue.use(Vuex)
-
-export default new Vuex.Store({
-  state: {
-  },
-  getters: {
-  },
-  mutations: {
-  },
-  actions: {
-  },
+export default createStore({
+  state: {},
+  mutations: {},
+  actions: {},
   modules: {
-  }
-})
+    auth,
+  },
+});
